refactor(daemon): split CC hooks from handler.go to cc_hooks.go (#182)
