+++ internal/module/agent/cc_hooks_test.go
@@ -1,0 +1,344 @@
+package agent
+
+import (
+	"encoding/json"
+	"net/http"
+	"net/http/httptest"
+	"os"
+	"path/filepath"
+	"strings"
+	"testing"
+)
+
+// writeSettingsJSON writes content into <dir>/.claude/settings.json and returns the file path.
+func writeSettingsJSON(t *testing.T, dir string, content string) string {
+	t.Helper()
+	claudeDir := filepath.Join(dir, ".claude")
+	if err := os.MkdirAll(claudeDir, 0755); err != nil {
+		t.Fatalf("mkdir .claude: %v", err)
+	}
+	p := filepath.Join(claudeDir, "settings.json")
+	if err := os.WriteFile(p, []byte(content), 0644); err != nil {
+		t.Fatalf("write settings.json: %v", err)
+	}
+	return p
+}
+
+// TestHandleHookStatus_NoSettingsFile — HOME points to a fresh temp dir
+// that has no .claude/settings.json; handler should report installed:false.
+func TestHandleHookStatus_NoSettingsFile(t *testing.T) {
+	m := newTestModule(t)
+
+	tmp := t.TempDir()
+	t.Setenv("HOME", tmp)
+
+	req := httptest.NewRequest("GET", "/api/hooks/cc/status", nil)
+	w := httptest.NewRecorder()
+	m.handleHookStatus(w, req)
+
+	if w.Code != http.StatusOK {
+		t.Fatalf("status: want 200, got %d (body: %s)", w.Code, w.Body.String())
+	}
+
+	var resp map[string]any
+	if err := json.NewDecoder(w.Body).Decode(&resp); err != nil {
+		t.Fatalf("decode response: %v", err)
+	}
+
+	if resp["installed"] != false {
+		t.Errorf("installed: want false, got %v", resp["installed"])
+	}
+
+	issues, _ := resp["issues"].([]any)
+	if len(issues) == 0 {
+		t.Fatal("issues should contain at least one entry")
+	}
+	found := false
+	for _, iss := range issues {
+		if s, ok := iss.(string); ok && strings.Contains(s, "settings.json not found") {
+			found = true
+			break
+		}
+	}
+	if !found {
+		t.Errorf("issues should mention 'settings.json not found', got %v", issues)
+	}
+}
+
+// TestHandleHookStatus_WithHooks — settings.json has a valid tbox hook command
+// for every expected event; handler should report installed:true.
+func TestHandleHookStatus_WithHooks(t *testing.T) {
+	m := newTestModule(t)
+
+	tmp := t.TempDir()
+	t.Setenv("HOME", tmp)
+
+	hookEvents := []string{"SessionStart", "UserPromptSubmit", "SubagentStart", "SubagentStop", "Stop", "StopFailure", "Notification", "PermissionRequest", "SessionEnd"}
+
+	// Build hooks map with a tbox hook entry for each event.
+	hooksMap := map[string]any{}
+	for _, ev := range hookEvents {
+		hooksMap[ev] = []any{
+			map[string]any{
+				"hooks": []any{
+					map[string]any{
+						"type":    "command",
+						"command": "/usr/local/bin/tbox hook " + ev,
+					},
+				},
+			},
+		}
+	}
+	settings := map[string]any{"hooks": hooksMap}
+	data, err := json.Marshal(settings)
+	if err != nil {
+		t.Fatalf("marshal settings: %v", err)
+	}
+	writeSettingsJSON(t, tmp, string(data))
+
+	req := httptest.NewRequest("GET", "/api/hooks/cc/status", nil)
+	w := httptest.NewRecorder()
+	m.handleHookStatus(w, req)
+
+	if w.Code != http.StatusOK {
+		t.Fatalf("status: want 200, got %d (body: %s)", w.Code, w.Body.String())
+	}
+
+	var resp map[string]any
+	if err := json.NewDecoder(w.Body).Decode(&resp); err != nil {
+		t.Fatalf("decode response: %v", err)
+	}
+
+	if resp["installed"] != true {
+		t.Errorf("installed: want true, got %v", resp["installed"])
+	}
+
+	issues, _ := resp["issues"].([]any)
+	if len(issues) != 0 {
+		t.Errorf("issues: want empty, got %v", issues)
+	}
+
+	events, ok := resp["events"].(map[string]any)
+	if !ok {
+		t.Fatal("events field missing or wrong type")
+	}
+	for _, ev := range hookEvents {
+		evData, ok := events[ev].(map[string]any)
+		if !ok {
+			t.Errorf("events[%s]: missing", ev)
+			continue
+		}
+		if evData["installed"] != true {
+			t.Errorf("events[%s].installed: want true, got %v", ev, evData["installed"])
+		}
+		cmd, _ := evData["command"].(string)
+		if !strings.Contains(cmd, "tbox hook") {
+			t.Errorf("events[%s].command: want to contain 'tbox hook', got %q", ev, cmd)
+		}
+	}
+}
+
+// TestHandleHookStatus_EmptyHooks — settings.json exists but hooks object is empty;
+// all events should be reported as not installed.
+func TestHandleHookStatus_EmptyHooks(t *testing.T) {
+	m := newTestModule(t)
+
+	tmp := t.TempDir()
+	t.Setenv("HOME", tmp)
+
+	writeSettingsJSON(t, tmp, `{"hooks":{}}`)
+
+	req := httptest.NewRequest("GET", "/api/hooks/cc/status", nil)
+	w := httptest.NewRecorder()
+	m.handleHookStatus(w, req)
+
+	if w.Code != http.StatusOK {
+		t.Fatalf("status: want 200, got %d (body: %s)", w.Code, w.Body.String())
+	}
+
+	var resp map[string]any
+	if err := json.NewDecoder(w.Body).Decode(&resp); err != nil {
+		t.Fatalf("decode response: %v", err)
+	}
+
+	if resp["installed"] != false {
+		t.Errorf("installed: want false, got %v", resp["installed"])
+	}
+
+	events, ok := resp["events"].(map[string]any)
+	if !ok {
+		t.Fatal("events field missing or wrong type")
+	}
+
+	hookEvents := []string{"SessionStart", "UserPromptSubmit", "SubagentStart", "SubagentStop", "Stop", "StopFailure", "Notification", "PermissionRequest", "SessionEnd"}
+	for _, ev := range hookEvents {
+		evData, ok := events[ev].(map[string]any)
+		if !ok {
+			t.Errorf("events[%s]: missing", ev)
+			continue
+		}
+		if evData["installed"] != false {
+			t.Errorf("events[%s].installed: want false, got %v", ev, evData["installed"])
+		}
+	}
+}
+
+// TestFindTboxCommand exercises the findTboxCommand helper directly.
+func TestFindTboxCommand(t *testing.T) {
+	tests := []struct {
+		name    string
+		entries any
+		want    string
+	}{
+		{
+			name:    "nil entries",
+			entries: nil,
+			want:    "",
+		},
+		{
+			name:    "not a slice",
+			entries: "string",
+			want:    "",
+		},
+		{
+			name:    "empty slice",
+			entries: []any{},
+			want:    "",
+		},
+		{
+			name: "entry without hooks key",
+			entries: []any{
+				map[string]any{"matcher": ".*"},
+			},
+			want: "",
+		},
+		{
+			name: "hooks list with no tbox command",
+			entries: []any{
+				map[string]any{
+					"hooks": []any{
+						map[string]any{"type": "command", "command": "/usr/bin/notify-send done"},
+					},
+				},
+			},
+			want: "",
+		},
+		{
+			name: "hooks list with tbox hook command",
+			entries: []any{
+				map[string]any{
+					"hooks": []any{
+						map[string]any{"type": "command", "command": "/opt/tbox hook Stop"},
+					},
+				},
+			},
+			want: "/opt/tbox hook Stop",
+		},
+		{
+			name: "multiple entries, tbox in second",
+			entries: []any{
+				map[string]any{
+					"hooks": []any{
+						map[string]any{"type": "command", "command": "echo hello"},
+					},
+				},
+				map[string]any{
+					"hooks": []any{
+						map[string]any{"type": "command", "command": "/usr/local/bin/tbox hook Notification"},
+					},
+				},
+			},
+			want: "/usr/local/bin/tbox hook Notification",
+		},
+		{
+			name: "quoted tbox path",
+			entries: []any{
+				map[string]any{
+					"hooks": []any{
+						map[string]any{"type": "command", "command": `"/Users/wake/Workspace/tmux-box/tbox" hook --agent cc Stop`},
+					},
+				},
+			},
+			want: `"/Users/wake/Workspace/tmux-box/tbox" hook --agent cc Stop`,
+		},
+		{
+			name: "hook map has no command field",
+			entries: []any{
+				map[string]any{
+					"hooks": []any{
+						map[string]any{"type": "command"},
+					},
+				},
+			},
+			want: "",
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			got := findTboxCommand(tc.entries)
+			if got != tc.want {
+				t.Errorf("findTboxCommand: want %q, got %q", tc.want, got)
+			}
+		})
+	}
+}
+
+// TestHandleHookSetup_BadJSON — invalid JSON body returns 400.
+func TestHandleHookSetup_BadJSON(t *testing.T) {
+	m := newTestModule(t)
+	req := httptest.NewRequest("POST", "/api/hooks/cc/setup", strings.NewReader("not json"))
+	req.Header.Set("Content-Type", "application/json")
+	w := httptest.NewRecorder()
+	m.handleHookSetup(w, req)
+	if w.Code != http.StatusBadRequest {
+		t.Fatalf("status: want 400, got %d (body: %s)", w.Code, w.Body.String())
+	}
+}
+
+// TestHandleHookSetup_InvalidAction — action other than install/remove returns 400.
+func TestHandleHookSetup_InvalidAction(t *testing.T) {
+	m := newTestModule(t)
+	req := httptest.NewRequest("POST", "/api/hooks/cc/setup", strings.NewReader(`{"action":"nuke"}`))
+	req.Header.Set("Content-Type", "application/json")
+	w := httptest.NewRecorder()
+	m.handleHookSetup(w, req)
+	if w.Code != http.StatusBadRequest {
+		t.Fatalf("status: want 400, got %d (body: %s)", w.Code, w.Body.String())
+	}
+}
+
+// TestHandleHookSetup_EmptyAction — empty action string returns 400.
+func TestHandleHookSetup_EmptyAction(t *testing.T) {
+	m := newTestModule(t)
+	req := httptest.NewRequest("POST", "/api/hooks/cc/setup", strings.NewReader(`{"action":""}`))
+	req.Header.Set("Content-Type", "application/json")
+	w := httptest.NewRecorder()
+	m.handleHookSetup(w, req)
+	if w.Code != http.StatusBadRequest {
+		t.Fatalf("status: want 400, got %d (body: %s)", w.Code, w.Body.String())
+	}
+}
+
+// TestHandleHookSetup_ExecFailure — valid action but tbox binary doesn't support
+// "setup" subcommand (running under go test), returns 500 with error detail.
+func TestHandleHookSetup_ExecFailure(t *testing.T) {
+	m := newTestModule(t)
+	req := httptest.NewRequest("POST", "/api/hooks/cc/setup", strings.NewReader(`{"action":"install"}`))
+	req.Header.Set("Content-Type", "application/json")
+	w := httptest.NewRecorder()
+	m.handleHookSetup(w, req)
+	if w.Code != http.StatusInternalServerError {
+		t.Fatalf("status: want 500, got %d (body: %s)", w.Code, w.Body.String())
+	}
+	var resp map[string]any
+	if err := json.NewDecoder(w.Body).Decode(&resp); err != nil {
+		t.Fatalf("decode: %v", err)
+	}
+	if resp["error"] != "setup failed" {
+		t.Errorf("error field: want 'setup failed', got %v", resp["error"])
+	}
+	if _, ok := resp["detail"]; !ok {
+		t.Error("response should contain 'detail' field with exec output")
+	}
+}
